Update OSX.md
I routinely ran into issues getting the right version of ruby to be 'seen' using RBENV, and as a beginner I was not aware of when, who, why, how, where RVM was installed. Nor was I aware how and why having RVM and RBENV installed simultaneously would pose an issue. So having this extra step to clear RVM in the event it is installed will save a lot of nubies hours of googling and frustration. - Or an argument could be made that I am just an imbecile