Mention that `sudo -u postgres` throws an error
Running ```sudo -u postgres psql --command "CREATE ROLE \"`whoami`\" LOGIN createdb superuser;"```
in the users home folder within ubuntu WSL throws the following error

```
could not change directory to "/home/student_name": Permission denied
```

The student doing the setup is expeceted to be in his /home/username folder.
Which on WSL ubuntu by default is not accessible to other users on the system.
So when the `su` command switches user, the postgres user can not access
the current directory and that is throwing the permission error.

Which is fine because it does not have to access the current folder at all.
It just has to connect to the postgres database which usually works.
So the permission warning is followed by the success output `CREATE ROLE`.

This might not be obvious to someone following the setup for the first time.
To reduce useless debugging there is now a note hidden in a spoiler that
says that this error is okay.