Fix Rails version for now (#473)
Seems like if students install rails 7.2.1, it creates issues with
sqlite3 and activerecord support, and I haven't been able to find a
solution for that (I'm pretty sure the latest versions of those gems
just aren't compatible). So, the best approach I can think of for now is
to fix our `rails` version. Also, to be perfectly honest, I don't love
the idea that periodic Rails upgrades can break our whole setup, so I
think this is also probably just a good idea anyway. But yeah, it is a
bit of a hack and will require us periodically upgrading the version
manually ourselves.

Also part of my wonders if I should upgrade the install command in that
first Rails lecture, but I don't think students ever end up needing to
run it, and it's after the DB unit anyway, so I think it should be fine.