Align Windows 11 and 10 WSL installation (#483)
* Align Windows 11 and 10 WSL installation

Starting from Windows 10 2004 (May 2020 Update), WSL 2 is the default
version of WSL. The instructions for installing WSL 2 are now the same
for both Windows 10 and Windows 11.

Adapted the instructions as such, to have a common installation process.

This will:
- make the setup easier for remaining Windows 10 users,
- reduce the  number of cases where Windows 11 users started following
  the Windows 10 instructions by mistake,
- make it easier to maintain the instructions in the future,
- make it easy to remove the Windows 10 instructions later.

For this we now run `wsl --install` in the Windows Command Prompt (cmd)
instead of Windows Terminal (wt, for Windows 11) or PowerShell (for
Windows 10).

For troubleshooting reasons, the instructions for Windows 10 < 2004 are
kept in details, just in case.

Adapted EN, FR, ES, PT.

* GHA: Build guides on PR

* GHA: Build guides on push on every branch

* setup guides generated

* Fine tune indentation for Windows 10

* setup guides generated

* Revert changes to build workflow

---------

Co-authored-by: github-actions <github-actions@github.com>