Set SSH default in gh auth command (#490)
* Set SSH default in gh auth command

Currently students have to select the protocol in the interactive
questions asked by `gh`. The first option is HTTP, and regularly is
chosen by accident instead of SSH. This remains undetected until late in
the setup, or even later, and then requires several steps (especially if
challenge repos have already been created).

This commit sets the correct protocol in the command, so it no longer
needs to be set in the interactive questions.

* Format gh questions as list