Add locale for ubuntu (#493)
* Correction typos in French

* Move locale from Win Ubuntu to CLI tools partial

The locale problem also happens on some Ubuntu installations.

The solution currently sits in the Windows Update partial, but it makes
more sense to include it in a partial that caters for both Windows and
Ubuntu.

The start of the CLI tools section is a good place to do this:
- It fits  better with the setup actions we do in that section.
- It is still early enough: the problem only gets triggered after dotfiles.

This commit will automatically move this in the right place. No need to
change the `build.rb`. The next rebuild of the data-setup will then also
include these changes. DA and DE setups do not rely on these partials.

Also corrected some typos in the FR partial.